refactor(trials): split admin trials into feature modules
- split the admin trials page, i18n, and server use-case into smaller feature-local modules
- add focused tests for the new split components and helpers while keeping the existing event-first flow intact

ref bej-91
