+++ apps/web/lib/i18n/messages/admin/trials/manage.ts
@@ -1,0 +1,168 @@
+export const fiAdminTrialsManageMessages = {
+  "admin.trials.title": "Ylläpito: ajokokeet",
+  "admin.trials.description":
+    "Selaa AJOK-kokeita tapahtumittain: valitse koe, tarkista koirat ja avaa detalji tai PDF.",
+  "admin.trials.manage.title": "Ajokokeiden koetulokset",
+  "admin.trials.manage.description":
+    "Hae ja selaa tapahtumia sekä niiden koirarivejä read-only-näkymässä.",
+  "admin.trials.manage.countSuffix": "tulosta",
+  "admin.trials.manage.filters.placeholder":
+    "Hae tapahtumaa paikalla, järjestäjällä, tuomarilla tai SKL-koeavaimella",
+  "admin.trials.manage.filters.aria": "Hae ajokoetapahtumia",
+  "admin.trials.manage.filters.sort.label": "Lajittelu",
+  "admin.trials.manage.filters.sort.dateDesc": "Uusin ensin",
+  "admin.trials.manage.filters.sort.dateAsc": "Vanhin ensin",
+  "admin.trials.manage.filters.mode.label": "Ajallinen rajaus",
+  "admin.trials.manage.filters.mode.year": "Vuosi",
+  "admin.trials.manage.filters.mode.range": "Päivämääräväli",
+  "admin.trials.manage.filters.year.placeholder": "Vuosi (esim. 2026)",
+  "admin.trials.manage.filters.year.label": "Vuosi",
+  "admin.trials.manage.filters.dateFrom": "Alkupäivä",
+  "admin.trials.manage.filters.dateTo": "Loppupäivä",
+  "admin.trials.manage.filters.apply": "Hae",
+  "admin.trials.manage.filters.reset": "Tyhjennä",
+  "admin.trials.manage.filters.validation.year":
+    "Vuosi pitää antaa muodossa YYYY.",
+  "admin.trials.manage.filters.validation.range":
+    "Valitse alku- ja loppupäivä.",
+  "admin.trials.manage.filters.validation.order":
+    "Alkupäivä ei voi olla loppupäivän jälkeen.",
+  "admin.trials.manage.loading": "Ladataan koetuloksia...",
+  "admin.trials.manage.error": "Koetulosten lataus epäonnistui.",
+  "admin.trials.manage.empty": "Ei koetuloksia näytettäväksi.",
+  "admin.trials.manage.events.title": "Tapahtumat",
+  "admin.trials.manage.events.subtitle":
+    "Valitse tapahtuma nähdäksesi siihen kuuluvat koirarivit.",
+  "admin.trials.manage.events.countSuffix": "tapahtumaa",
+  "admin.trials.manage.events.empty": "Ei tapahtumia valituilla ehdoilla.",
+  "admin.trials.manage.events.columns.date": "Päivä",
+  "admin.trials.manage.events.columns.place": "Paikka",
+  "admin.trials.manage.events.columns.name": "Tapahtuma",
+  "admin.trials.manage.events.columns.judge": "Ylituomari",
+  "admin.trials.manage.events.columns.dogs": "Koiria",
+  "admin.trials.manage.selected.title": "Valittu tapahtuma",
+  "admin.trials.manage.selected.subtitle":
+    "Tapahtuman koirarivit ja rivikohtaiset toiminnot.",
+  "admin.trials.manage.selected.loading": "Ladataan tapahtuman koirarivejä...",
+  "admin.trials.manage.selected.error":
+    "Tapahtuman koerarivien lataus epäonnistui.",
+  "admin.trials.manage.selected.empty":
+    "Valitse ensin tapahtuma tapahtumalistalta.",
+  "admin.trials.manage.selected.countSuffix": "koiraa",
+  "admin.trials.manage.selected.emptyRows":
+    "Valitussa tapahtumassa ei ole koerarivejä.",
+  "admin.trials.manage.selected.columns.dog": "Koira",
+  "admin.trials.manage.selected.columns.registration": "Rekisterinumero",
+  "admin.trials.manage.selected.columns.points": "Pisteet",
+  "admin.trials.manage.selected.columns.award": "Palkinto",
+  "admin.trials.manage.selected.columns.rank": "Sija",
+  "admin.trials.manage.selected.columns.judge": "Tuomari",
+  "admin.trials.manage.selected.columns.actions": "Toiminnot",
+  "admin.trials.manage.selected.actions.openDetail": "Avaa detalji",
+  "admin.trials.manage.selected.actions.openPdf": "Avaa PDF",
+  "admin.trials.manage.pagination.previous": "Edellinen",
+  "admin.trials.manage.pagination.page": "Sivu",
+  "admin.trials.manage.pagination.next": "Seuraava",
+  "admin.trials.manage.columns.dog": "Koira",
+  "admin.trials.manage.columns.registration": "Rekisteri / SKL-koeavain",
+  "admin.trials.manage.columns.date": "Päivä",
+  "admin.trials.manage.columns.place": "Paikka",
+  "admin.trials.manage.columns.piste": "Pisteet",
+  "admin.trials.manage.columns.pa": "Pa",
+  "admin.trials.manage.columns.sija": "Sija",
+  "admin.trials.manage.columns.judge": "Tuomari",
+  "admin.trials.manage.results.openDetailHint":
+    "Avaa detalji napauttamalla riviä.",
+  "admin.trials.manage.results.openDetailAriaPrefix":
+    "Avaa koetuloksen detalji",
+  "admin.trials.manage.mobile.date": "Päivä",
+  "admin.trials.manage.mobile.place": "Paikka",
+  "admin.trials.manage.mobile.piste": "Pisteet",
+  "admin.trials.manage.mobile.pa": "Pa",
+  "admin.trials.manage.mobile.sija": "Sija",
+  "admin.trials.manage.mobile.judge": "Tuomari",
+} as const;
+
+export const svAdminTrialsManageMessages = {
+  "admin.trials.title": "Admin: jaktprov",
+  "admin.trials.description":
+    "Bläddra AJOK-prov per evenemang: välj prov, granska hundrader och öppna detalj eller PDF.",
+  "admin.trials.manage.title": "Jaktprovresultat",
+  "admin.trials.manage.description":
+    "Sök och bläddra evenemang och deras hundrader i en read-only-vy.",
+  "admin.trials.manage.countSuffix": "resultat",
+  "admin.trials.manage.filters.placeholder":
+    "Sök evenemang via plats, arrangör, domare eller SKL-provnyckel",
+  "admin.trials.manage.filters.aria": "Sök jaktprovsevenemang",
+  "admin.trials.manage.filters.sort.label": "Sortering",
+  "admin.trials.manage.filters.sort.dateDesc": "Nyaste först",
+  "admin.trials.manage.filters.sort.dateAsc": "Äldsta först",
+  "admin.trials.manage.filters.mode.label": "Tidsfilter",
+  "admin.trials.manage.filters.mode.year": "År",
+  "admin.trials.manage.filters.mode.range": "Datumintervall",
+  "admin.trials.manage.filters.year.placeholder": "År (t.ex. 2026)",
+  "admin.trials.manage.filters.year.label": "År",
+  "admin.trials.manage.filters.dateFrom": "Startdatum",
+  "admin.trials.manage.filters.dateTo": "Slutdatum",
+  "admin.trials.manage.filters.apply": "Sök",
+  "admin.trials.manage.filters.reset": "Rensa",
+  "admin.trials.manage.filters.validation.year":
+    "Året måste anges i formatet YYYY.",
+  "admin.trials.manage.filters.validation.range":
+    "Välj både start- och slutdatum.",
+  "admin.trials.manage.filters.validation.order":
+    "Startdatum kan inte vara efter slutdatum.",
+  "admin.trials.manage.loading": "Laddar provresultat...",
+  "admin.trials.manage.error": "Det gick inte att ladda provresultat.",
+  "admin.trials.manage.empty": "Inga provresultat att visa.",
+  "admin.trials.manage.events.title": "Evenemang",
+  "admin.trials.manage.events.subtitle":
+    "Välj ett evenemang för att se tillhörande hundrader.",
+  "admin.trials.manage.events.countSuffix": "evenemang",
+  "admin.trials.manage.events.empty": "Inga evenemang med valda filter.",
+  "admin.trials.manage.events.columns.date": "Datum",
+  "admin.trials.manage.events.columns.place": "Plats",
+  "admin.trials.manage.events.columns.name": "Evenemang",
+  "admin.trials.manage.events.columns.judge": "Överdomare",
+  "admin.trials.manage.events.columns.dogs": "Hundar",
+  "admin.trials.manage.selected.title": "Valt evenemang",
+  "admin.trials.manage.selected.subtitle":
+    "Hundrader och radåtgärder för valt evenemang.",
+  "admin.trials.manage.selected.loading": "Laddar hundrader för evenemanget...",
+  "admin.trials.manage.selected.error":
+    "Det gick inte att ladda evenemangets hundrader.",
+  "admin.trials.manage.selected.empty": "Välj först ett evenemang i listan.",
+  "admin.trials.manage.selected.countSuffix": "hundar",
+  "admin.trials.manage.selected.emptyRows":
+    "Det finns inga hundrader i valt evenemang.",
+  "admin.trials.manage.selected.columns.dog": "Hund",
+  "admin.trials.manage.selected.columns.registration": "Registrering",
+  "admin.trials.manage.selected.columns.points": "Poäng",
+  "admin.trials.manage.selected.columns.award": "Pris",
+  "admin.trials.manage.selected.columns.rank": "Placering",
+  "admin.trials.manage.selected.columns.judge": "Domare",
+  "admin.trials.manage.selected.columns.actions": "Åtgärder",
+  "admin.trials.manage.selected.actions.openDetail": "Öppna detalj",
+  "admin.trials.manage.selected.actions.openPdf": "Öppna PDF",
+  "admin.trials.manage.pagination.previous": "Föregående",
+  "admin.trials.manage.pagination.page": "Sida",
+  "admin.trials.manage.pagination.next": "Nästa",
+  "admin.trials.manage.columns.dog": "Hund",
+  "admin.trials.manage.columns.registration": "Reg / SKL-provnyckel",
+  "admin.trials.manage.columns.date": "Datum",
+  "admin.trials.manage.columns.place": "Plats",
+  "admin.trials.manage.columns.piste": "Poäng",
+  "admin.trials.manage.columns.pa": "Pa",
+  "admin.trials.manage.columns.sija": "Placering",
+  "admin.trials.manage.columns.judge": "Domare",
+  "admin.trials.manage.results.openDetailHint":
+    "Öppna detaljen genom att trycka på raden.",
+  "admin.trials.manage.results.openDetailAriaPrefix":
+    "Öppna provresultatdetalj",
+  "admin.trials.manage.mobile.date": "Datum",
+  "admin.trials.manage.mobile.place": "Plats",
+  "admin.trials.manage.mobile.piste": "Poäng",
+  "admin.trials.manage.mobile.pa": "Pa",
+  "admin.trials.manage.mobile.sija": "Placering",
+  "admin.trials.manage.mobile.judge": "Domare",
+} as const;
